pyproject.toml
[project]
name = "music-store-support-bot"
version = "0.1.0"
description = "Customer support chatbot for a music store using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    # External API integrations
    "google-api-python-client>=2.0.0",  # YouTube Data API
    "twilio>=8.0.0",                     # Twilio Verify API
    "requests>=2.28.0",                  # Genius API
]

[project.scripts]
music-support = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

